feat(sdk): refactor chain constants
Refactors contract addresses and related constants into the
chain-constants.ts file. Much cleaner separation of logic.
