+++ .changeset/real-ravens-clean.md
@@ -1,0 +1,5 @@
+---
+'@eth-optimism/sdk': patch
+---
+
+Minor refactor to variables within the SDK package.
